Two Apps - Homescreen and basic simon (version 1)
simon_project starts the whole sequence. if you click "play simon 1.0" the second app opens, which is the base version of the game.

simon_1 is the current progress (version 1) on the game. It initializes the game board, creates the random game sequence of color patterns, and  initiates a blank user array.
On the first turn, it is set to (poorly) highlight the button set by the sequence and (too quickly) plays the sound of that button. 

Further work is needed to make the button flashing better. When the commented-out commands (that set visibile to off) are active, the flashing of the button is too fast and the users can't see it. Ideally, the background OG_simon_game board can be changed from just an image (what it is now..) and coded up in the App itself. This way, when the buttons are clicked they can remain invisible and the colored area can be coded to simply change color.. This will have to be accompanied by a timer or time limit, which I have yet to look up..